% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eClassic.R
\docType{package}
\name{eClassic-package}
\alias{eClassic}
\alias{eClassic-package}
\title{eClassic: Efficient Classic Financial Factors Toolkit}
\description{
A fast and lightweight package for computing classic quantitative
factors such as size, value, momentum, profitability, quality,
volatility, and liquidity.
}
\author{
Yishuo Deng \href{mailto:dengyishuo@163.com}{dengyishuo@163.com}
}
\keyword{internal}
